examples/requirements.txt
# optional dependencies for cli.py (pretty output / YAML / alternate progress)
rich>=13
tabulate>=0.9
PyYAML>=6
tqdm>=4.65

# for examples/icons.py
Pillow>=9.1